Replace Leaf2 with Leaf in `node-metrics`
The node validator API consumes from endpoint `/v0/` which **should** only return
`Leaf` instead of `Leaf2`.  Because of this, the current implementation will always fail
to decode a `Leaf2` due to the serialization being incompatible.

Additionally, instead of producing `Leaf2`, or `Leaf` the `stream_leaves` endpoint
instead seems to stream `LeafQueryData`.  This has lead to some further
incompatability in the API.  `Leaf` seems to have worked historically, but now it
no longer can.

This commit replaces `Leaf2` with `Leaf1QueryData`, and updates the derived logic.
Additionally, it adds some necessary trait implementations back to `Leaf1QueryData`
so that it is able to access what it needs to.

Remove CDN related tasks

The CDN consumption components are not intended to ever be used, and the RollCall
functionality has been removed from the sequencer.  The node validator is also never
meant to be permissioned, so this makes it ineligible to received messages directly
from the CDN.