Clean up redundant code in `test_full_setup_example`
The initial code in `test_full_setup_example` was setup and existed before the full
implementation of the `run_standalone_service` with the `Options` structure.  As the
project has evolved, the setup of the the various configuration options have been fleshed
out and implemented, making this example perform a lot of unnecessary work.

This change cleans up this unnecessary code in favor of the existing basic
implementations.