CI: use less caches
Our current cache usage is 100GB (10x what github claims to allow). As a
result our caches are evicted often and on `main` we sometimes don't
have a live cache anymore.

This PR changes all workflows to save the cache from main only. In the
hopes of reducing usage. It also shares the cache for job pairs where we
compile without and with the sqlite feature.