fix: demo sometimes deploys contracts to early
Using netcat to wait for the socket being open is not sufficient. We
should wait until the RPC is actually responding.

Also noticed that by default the demo-native recipe builds the binaries
with the dev profile which is known to be very wasteful and cause
issues. This changes the default to the "test" profile which works much
better due to having some optimizations turned on.