use nodejs base image
This should be a fairly quick build since the gazillion packages needed
to install npm are installed in the base image and not in the CI job
that runs in this repo.

Overall the deploy image size is reduced from about 900MB to 550MB.
Still pretty large but also considerably less than before.