Add support for subscribed stake table and validator updates
When the `DataState` is processing blocks, it will encounter an epoch
transition.  When this occurs it will update it's local `StakeTable`, and
`Nodes` list in order  to ensure that the Voters bitvec is reflecting the
local Node Identity State, rather than the Stake Table.  At this time we
also update the Validator list for the stake table as their their stake and
delegations may have changed.

This serves as a prime location to send the Stake Table and Validator
updates to any clients that have subscribed to them.

This commit adds support to the Subscribe Client Message requests
by ensuring that the updated Stake Table, and the Validator changes
are sent to the client.

Fix too many arguments lint error

With the increase in the number of Sender Sinks being passed to
the `ProcessLeafAndBlockPairStreamTask` task, we ended up running
against the lint for having too many arguments.  To address this, it
could be beneficial to bundle the Senders together in a tuple.

This commit bundles the senders send to
`ProcessLeafAndBlockPairStreamTask` into a tuple.

Fix lint calling `clone` on `Copy`

The `Address` type that is used for the validators implements `Copy`
meaning that it doesn't need to be cloned directly.  As a result any
call to `clone` on this address should be dropped.

This commit drops the `clone` call from the `address`.

Fix lint: let_and_return

In the `handle_send_server_message_to_subscribed_clients` that has
been added to reduce the amount of repeated code in the
`node-metrics` there is an unnecessary `let` statement right before
the value is returned.  This `let` statement can be dropped.

This commit simply returns the calculated value instead of storing it
locally initially.

Fix remaining clippy errors