Remove PublicHotShotConfig in favor of existing definition
@imabdulbasit suggested that the public type `PublicHotShotConfig` be
utilized instead of re-creating the type locally for the package.  This is
primarily due to the type changing and no longer being compatible with
the existing type definition.  (Option was removed from the numeric
values for the epoch configuration).

This commit removes the local definition of `PublicHotShotConfig` in
favor of the pre-existing definition.