Add unrecognized request case to node-metrics
As we transition the node validator dashboard and block explorer to
handling more request types, we may find ourselves in the situation
where we have a mismatch in the UI version and the Server version.
The most egregious mismatch of these cases would be where the
Client UI has a deployed version ahead of the Server's deployed
version.

In the event that the Client UI version is ahead of the Node Metrics's
version API, we may end up in a scenario where the Client is sending
a request that would be valid on the new server, but is currently invalid
on the existing server.  In this specific case, due to the nature of how
`tide_disco` handles client messages, this error would not be detected
or logged on the server, and would result in the client being
disconnected without a given reason.  This would not be detectably
different from a network disconnect issue as far as the client is
concerned.

As a result, there should be a catch-all case that handles these invalid
inputs received.

NOTE: there is still a case for the client to be disconnected without
reason, namely that the message sent to the server isn't in the expected
format / content-type.  In this case `tide_disco` will still disconnect the
user silently.

This commit adds invalid message handling to the node-metrics program
to inform the user about the invalid message, and to keep processing the
other requests the user may have.