Modify config structs for node-metrics for compatibility
The `PublicHotShotConfig` struct that is in use by the `node-metrics`
was originally created to only define what is actively being consumed.
It was modified to use the pre-existing types in favor of ease-of-use
and correctness.  However, when attempting to run against `mainnet`
it was discovered that the `espresso_types` definitions of
`PublicHotShotConfig` isn't currently compatible with what is running
on mainnet.

In order to ensure this compatibility, this commit reverts to defining the
structure locally with a minimum set of values being retrieved.
Additionally, a test has been added to ensure compatibility with the
`espresso_types` defined serialization.