Modify sdk management script to remove potential security vulnerabilities (#3485)
Rename download binary to cmd/espresso-network-lib-utility

The folder `download` under `sdks/go` is meant to result in an executable binary that can
be run to help manage the espresso sdk library file.  It's name is not very descriptive
or indicative of what it is attempting to do / achieve.  Additionally, it is not under
a `cmd` folder, which would help to indicate that it is meant to be an entrypoint for
code evaluation.

This renames the file to a more descriptive name, and relocates the file under a
`cmd` folder to indicate its intent.

Modify script to remove special behavior

This script is expected to be run against the remote version of this repository in order
to determine the correct directory to symbolically link the downloaded binary into.
This is problematic as it means running code from a repo that may change without
warning, and could conceivably be a target for corruption.

To address this, we derive the path of the go mod cache utilizing and respecting
environment variables to overwrite them.  Additionally we utilize
`golang.org/x/mod/module` to help us properly escape and build the path so it
corresponds to the actual implementation rather than our understood representation.

Fix the need for this script to be run with elevated permissions

The execution of this script often relies on elevating the permissions of the user
running it using `sudo`.  The main reason for this has to do with the permission that
golang sets on the files it manages in the go module cache.  It makes them read-only,
but still owned by the user.

To address this, we take advantage of the fact that we can change the permissions of
files we own.  So we will modify the folders as needed, and revert them when we
are done.  Additionally, the cleanup script has been modified as it doesn't need to
recursively remove the folder and all of its contents.