fix: approval txns fail with geth in demo
For some reason the behaviour is different than anvil. I'm not sure if
this always fails due to how geth works or if it's just a race
condition. In any event it seems adding another transaction receipt
await point for approvals seems easier than getting to the bottom of
this.

If this turns out to be too slow we should just manually track nonces.