WIP: remove process compose tech debt
The new process compose handles the environment differently.

If an env var is set in .env it can't be overridden in the
`environment:` sections. Or maybe I can't figure out how.

So we remove all env vars that we might want to override later
out of the `.env` file and handle them in process-compose.yaml
or docker-compose.yaml instead.