Streamline light client leaf API
Much of the data in a chain of `LeafQueryData` is redundant (e.g.
the `qc()` of one leaf is the same as the `justify_qc()` of the
next). All we really need is a chain of the raw `Leaf2`s. A
`LeafQueryData` can be reconstructed on the client side by combining
the first leaf in the chain with the `justify_qc` of the next.

Also pulls out the last 2 QCs, since we don't need the leaf part
of those, which simplifies verification since the 2-chain of QCs
is always in the proof's `qc`s field (not sometimes in the end of
the leaf chain and other times in the `extra_qcs` field).