feat: staking-cli: concurrent txns, better output (#3639)
* feat: concurrent txns, better output

- allow broadcasting txns without awaiting
- output events
- check for events in tests
- add test to inspect output manually
- output to stdout/stderr consistently
- always use tracing if RUST_LOG_FORMAT=json

* fix: don't use deprecated method

* test: demo deploy without instant mining

(currently fails)

* fix: wait for dependent txns to finalize

* stake for demo with planned transactions

- Add a type to represent planned transactions.
- Give caller control over submission and awaiting of transactions
  while enforcing that dependencies are respected.
- Add balance check.
- Add tests.

* add missing "event" prefix

* fix: approval txns fail with geth in demo

For some reason the behaviour is different than anvil. I'm not sure if
this always fails due to how geth works or if it's just a race
condition. In any event it seems adding another transaction receipt
await point for approvals seems easier than getting to the bottom of
this.

If this turns out to be too slow we should just manually track nonces.

* fix: nonce errors due to duplicate provider

* remove unnecessary pending block ID

* cleanup: centralize receipt status check

* remove some useless variables

* update docs

* simplify types

- Remove provider from transaction enum variants
- Remove token and stake table addresses from transaction enum variants
- Consume transaction inputs when creating transactions

* separate processor and data types

This works much better with ownership. Make code cleaner and we don't
have to use `take` anymore.