Add an integration-style test for light client leaf validation
The test runs a local testnet through a protocol upgrade and epoch
change, then verifies that it can use the light client to fetch and
validate leaves throughout the network's history, before and after
each important event.

Also makes some tweaks to the existing `test_upgrade_helper` which
I think improve the robustness and debuggability of that test (which
was used as a model for the new test).