Implement light client leaf endpoint and validation (#3680)
* Implement light client leaf endpoint

* Streamline light client leaf API

Much of the data in a chain of `LeafQueryData` is redundant (e.g.
the `qc()` of one leaf is the same as the `justify_qc()` of the
next). All we really need is a chain of the raw `Leaf2`s. A
`LeafQueryData` can be reconstructed on the client side by combining
the first leaf in the chain with the `justify_qc` of the next.

Also pulls out the last 2 QCs, since we don't need the leaf part
of those, which simplifies verification since the 2-chain of QCs
is always in the proof's `qc`s field (not sometimes in the end of
the leaf chain and other times in the `extra_qcs` field).

* Move light client types to their own crate

This will allow the actual light client service to use these types
without importing the entire Espresso sequencer crate

* Implement QC chain validation, handling upgrades and signature checks

* Handle epoch changes in light client

* Expose next epoch justify QCs in decide event
* Store next epoch justify QCs in consensus ephemeral storage, so
  we can reconstruct decide events when triggered
* Validate next epoch justify QCs when they are required to be
  present in the light client

* Remove confusing/outdated comments

* Add sanity checks for server-side invariants

* Add an integration-style test for light client leaf validation

The test runs a local testnet through a protocol upgrade and epoch
change, then verifies that it can use the light client to fetch and
validate leaves throughout the network's history, before and after
each important event.

Also makes some tweaks to the existing `test_upgrade_helper` which
I think improve the robustness and debuggability of that test (which
was used as a model for the new test).

* Wait for second epoch change in light client completeness test

* Enforce finalized is after requested leaf in light client API

* Refactor validate_qc_and_next_epoch_qc to use CertificatePair

* trace -> debug