Handle validator entry and exit events (#12)
* Decouple L1 state data types from public API types

* Some of the info in the API types is redundant. E.g. each
  individual wallet and nodes snapshot contains L1 block info. But
  in the L1 state, we already have this information in the top-level
  `BlockData` structure
* We may want to store additional information in the backend that is
  not exposed to the API, or store information in a different structure.
  For example, the backend wants nodes stored in a persistent map indexed
  by address, for efficient updates and snapshotting. The frontend doesn't
  care about this and just wants a `Vec` for simplicity

* Implement node registration event handling

* Make event handling stateful

This is necessary in various cases. Most immediately, processing the
`Deregister` event requires access to the previous state, because we
are supposed to produce a `FullNodeSetDiff::NodeExit`, which requires
us to know the exit time, which requires us to know the
`exit_escrow_period`, which in turn depends on previous state and
events (since it can be changed causing a contract event to be emitted).

This statefulness may come in handy in other ways, too, for example if
we want to enforce uniqueness of keys in the service rather than rely
on the contract to do it, then event processing will need to know the
full set of keys in the previous state.

This is a pretty big refactor in terms of LOC but it really shouldn't
change much logic, it's just a matter of plumbing.

* Switch from `StakeTableEvent` to raw contract events

This allows us to handle additional events which are relevant to
this service but not consensus, such as `ExitEscrowPeriodUpdated`

* Test event handling

* Patch HotShot

* Fix test

* Remove `EpochAndBlock` field from L1 events

It's not clear what this field would be used for, and populating it
is surprisingly hard given that the events are generated by the L1
while the epoch is a property of Espresso.

After discussing with Theo we decided it's fine to delete these
fields for now.

* Switch back to main HotShot