fix: check timestamp drift w.r.t. validation start
We should validate that the block header timestamp does not drift too
much from the system time at the time we start validating the header.
Before this change the catchup is performed before the timestamp drift
validation. Since catchup can be slow this may cause the validation to
fail. It would be better to separate the quick "stateless" validation
and do that before applying the header but that's a major refactor to
core logic so isn't done here.