Break demo prerequisites into separate CI steps (#20)
* Break demo prerequisites into separate CI steps

On one branch, it looked like the CI demo smoketest timed out while
Docker images were still being pulled. This is because we had a
single workflow step responsible for pulling Docker images, starting
the demo, and waiting for things to become healthy. This whole step
ran in the background so we were starting the smoketest immediately.

This change breaks Docker pulling, and waiting for background
services to become healthy, into their own steps. We wait for these
to complete before proceeding to next steps.

* Wait for services to become healthy before just demo::up returns