Implement active node endpoints and storage (#27)
* Update demo instructions

* Implement active node endpoints

The crux of the change here is pretty small (mostly just app.rs).
I had to do a bunch of seemingly unrelated stuff to get the demo
to work nicely with Espresso, namely:
* Make the Espresso state wait for the first epoch rather than
  panicking if started before the first epoch
* To avoid that taking forever, make Espresso epochs very fast in
  the docker-compose
* To avoid Espresso entering the first epoch before the stake table
  transactions finalize, make L1 epochs also really fast (which
  requires switching from Geth to Anvil, to be able to control this)
* Build the executable with `--testing` for the time being, because
  I am instantiating it with the test-only `MemoryStorage` for now.
  That will change once we implement persistence for Espresso state

* Active node set persistence

* Add indexed active nodes endpoint for better cacheability

* Update env vars docs

* Fix demo test in CI

* Fix a rare edge case that can happen if the active node set decreases between epochs

* Make active node set updates idempotent.

This works around an issue we encountered in Decaf where AWS may
briefly run two copies of the service at once. If we're not
careful, this can lead to double counting votes and such.